AUT-1207 - Rename verify-mfa-code-post to verify-mfa-code-controller
